requirements.txt
streamlit>=1.34.0
requests
pymysql
python-dotenv
